Restructure folders and files
Renamed folders and moved files.
Updated CMakeList.txt files because of these changes.
