+++ klvp/include/klvp/encode.h
