+++ klvp/include/klvp/util.h
