+++ klvp/src/KlvSizeVisitor.cpp
@@ -1,6 +1,5 @@
-#include "KlvSizeVisitor.h"
-
-#include "util.h"
+#include <klvp/KlvSizeVisitor.h>
+#include <klvp/util.h>
 
 namespace
 {
