+++ klvp/src/encode.cpp
@@ -1,6 +1,5 @@
-#include "encode.h"
-
-#include "FpParser.h"
+#include <klvp/encode.h>
+#include <klvp/FpParser.h>
 
 #include <string.h>
 #include <math.h>
