+++ klvp/src/klvelmt.cpp
@@ -1,8 +1,6 @@
-#include "klvelmt.h"
-
-#include "klvelmtimpl.h"
-#include "klvuselmtimpl.h"
-
+#include <klvp/klvelmt.h>
+#include <klvp/klvelmtimpl.h>
+#include <klvp/klvuselmtimpl.h>
 
 lcss::KLVElement::KLVElement()
 {
