+++ ldsdb/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required (VERSION 3.14)
-project("LibLdsDB"
-    VERSION 1.0 
+project(ldsdb
+    VERSION 0.1.0 
     DESCRIPTION "STANAG 4609 ST 0601 LDS Database library"
     LANGUAGES C CXX)
 
@@ -15,31 +15,33 @@ else()
 	include_directories(../../sqlite3-cmake/libs/include)
 endif()
 
-file(GLOB SRC_LIST
-    *.h
-    *.cpp)
-add_library(LibLdsDB STATIC ${SRC_LIST})
-set_property(TARGET LibLdsDB PROPERTY POSITION_INDEPENDENT_CODE ON)
+add_library(ldsdb STATIC)
 
-target_include_directories(LibLdsDB
-          PUBLIC 
-          "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>"
-          "$<INSTALL_INTERFACE:include>"
+file(GLOB SRC_LIST
+    src/*.cpp)
+	
+target_sources(ldsdb PRIVATE ${SRC_LIST})
+
+set_property(TARGET ldsdb PROPERTY POSITION_INDEPENDENT_CODE ON)
+
+target_include_directories(ldsdb
+          PRIVATE src/ldsdb
+		  PUBLIC include
           )
 
-file(GLOB LIB_INCLUDE_FILES "./*.h")
+#file(GLOB LIB_INCLUDE_FILES "./*.h")
 
 # install the target and create export-set
-install(TARGETS LibLdsDB 
-    EXPORT LibLdsDBTargets
-    LIBRARY DESTINATION lib
-    ARCHIVE DESTINATION lib
-    RUNTIME DESTINATION bin
-    INCLUDES DESTINATION include
-)
+#install(TARGETS LibLdsDB 
+#    EXPORT LibLdsDBTargets
+#    LIBRARY DESTINATION lib
+#    ARCHIVE DESTINATION lib
+#    RUNTIME DESTINATION bin
+#    INCLUDES DESTINATION include
+#)
 
 # install header file
-install(FILES ${LIB_INCLUDE_FILES} DESTINATION include)
+#install(FILES ${LIB_INCLUDE_FILES} DESTINATION include)
 
 # generate and install export file
 #install(EXPORT LibLdsDBTargets
