+++ src/TestKLVParser.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "klvprsr.h"
+#include <klvp/klvprsr.h>
+
 #include "KLVPrintVisitor.h"
 
 class TestKLVParser :
