+++ src/TestKLVSecuritySetParser.cpp
@@ -1,7 +1,7 @@
 #include "TestKLVSecuritySetParser.h"
 
-#include "klvelmt.h"
-#include "klvelmtimpl.h"
+#include <klvp/klvelmt.h>
+#include <klvp/klvelmtimpl.h>
 
 #include <iostream>
