+++ src/TestKLVSecuritySetParser.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "klvprsr.h"
+#include <klvp/klvprsr.h>
+
 #include "KLVSecuritySetPrintVisitor.h"
 
 class TestKLVSecuritySetParser :
